+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # Summary
 * [**Unreleased**](#unreleased)
+* [**Release 2025-03-05 v0.3a1**](#release-2025-03-05-v03a1) 
 * [**Release 2025-02-25 v0.3a0**](#release-2025-02-25-v03a0) 
 * **Release 2025-02-18 v0.2a5** Changed logo to a stopwatch
 * **Release 2025-02-18 v0.2a4** Added `trk time` command to report elapsed hours
@@ -18,6 +19,10 @@
 # Details
 
 ## Unreleased
+
+## Release 2025-03-05 v0.3a1
+* ADD report command
+  * Initial implementation works for a single user on a single project (more to come)
 
 ## Release 2025-02-25 v0.3a0
 * ADD `--at` and `--force` to `start` and `stop` commands